+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (world module) fix: should not stuck the server if some `datapack` use `scheduled functions`. ([#263](https://github.com/sakurawald/fuji/issues/263]) Thanks to @wolgamir)
+- (command.spy module) feature: add `only_spy_these_commands` option. ([#266](https://github.com/sakurawald/fuji/issues/266) Thanks to @Steedr)